Allow Armada to clone OpenStack infra repos
-No longer attempts to ping repo
-Move all source locating (git or local) to preflight
-Only unique repos are cloned now
-Catches any source/url errors in preflight and raises exception

+++ armada/utils/git.py
@@ -1,16 +1,14 @@
 import pygit2
 import tempfile
 import shutil
-import requests
-
-HTTP_OK = 200
+from os import path
 
 def git_clone(repo_url, branch='master'):
     '''
     clones repo to a /tmp/ dir
     '''
 
-    if repo_url == '' or not check_available_repo(repo_url):
+    if repo_url == '':
         return False
 
     _tmp_dir = tempfile.mkdtemp(prefix='armada', dir='/tmp')
@@ -22,15 +20,5 @@ def source_cleanup(target_dir):
     '''
     Clean up source
     '''
-    shutil.rmtree(target_dir)
-
-
-def check_available_repo(repo_url):
-    try:
-        resp = requests.get(repo_url.replace('git:', 'http:'))
-        if resp.status_code == HTTP_OK:
-            return True
-
-        return False
-    except Exception:
-        return False
+    if path.exists(target_dir):
+        shutil.rmtree(target_dir)
